UBLE-176 refactor: 훅 제거
